+++ Scene.h
@@ -9,9 +9,9 @@ namespace Manbat {
 			Vector3 PlayerStart;
 			ImprovedCamera* playerController;
 			void sortBillboardsByDistance() {
-				for (int i = 0; i < contents.size(); i++) {
+				for (size_t i = 0; i < contents.size(); i++) {
 					if (contents[i]->getEntityType() == ENTITY_BILLBOARD) {
-						for (int j = 0; j < contents.size(); j++) {
+						for (size_t j = 0; j < contents.size(); j++) {
 							if (contents[j]->getEntityType() == ENTITY_BILLBOARD) {
 								if (Math::Distance(playerController->camera->getPosition(), contents[j]->getPosition()) < Math::Distance(playerController->camera->getPosition(), contents[i]->getPosition())) {
 									ImprovedMesh* temp = contents[i];
